requirements.txt
openai>=1.0.0
python-dotenv>=1.0.0
pymupdf==1.23.8
pdfplumber==0.10.3
Pillow>=10.0.0
rich>=13.0.0
jinja2>=3.0.0
markdown>=3.5.0
numpy>=1.26.0